+++ CHANGES.rst
@@ -25,6 +25,25 @@ Upcoming 2.0 Changes
 
 * Removed ``BodyNotHttplibCompatible`` and ``ResponseNotChunked`` exceptions.
 
+dev (master)
+------------
+
+* Change ``HTTPSConnection`` to load system CA certificates
+  when ``ca_certs``, ``ca_cert_dir``, and ``ssl_context`` are
+  unspecified. (Pull #1608, Issue #1603)
+
+* Upgrade bundled rfc3986 to 1.3.2. (Pull #1609, Issue #1605)
+
+
+1.25.2 (2019-04-28)
+-------------------
+
+* Change ``is_ipaddress`` to not detect IPvFuture addresses. (Pull #1583)
+
+* Change ``parse_url`` to percent-encode invalid characters within the
+  path, query, and target components. (Pull #1586)
+
+
 1.25.1 (2019-04-24)
 -------------------
 
@@ -60,6 +79,10 @@ Upcoming 2.0 Changes
 
 * Implemented a more efficient ``HTTPResponse.__iter__()`` method. (Issue #1483)
 
+1.24.3 (2019-05-01)
+-------------------
+
+* Apply fix for CVE-2019-9740. (Pull #1591)
 
 1.24.2 (2019-04-17)
 -------------------